+++ youtube_audio_matcher/audio/__init__.py
@@ -1,11 +1,13 @@
-from .audio import (
+from .fingerprint import (
     find_peaks_2d, fingerprint_from_file, fingerprint_from_signal,
-    get_spectrogram, hash_file, hash_peaks, plot_peaks, plot_spectrogram,
-    read_file,
+    get_spectrogram, hash_peaks, plot_peaks, plot_spectrogram
 )
+
+from .util import generate_waveform, hash_file, read_file
+
 
 __all__ = [
     "find_peaks_2d", "fingerprint_from_file", "fingerprint_from_signal",
-    "get_spectrogram", "hash_file", "hash_peaks", "plot_peaks",
-    "plot_spectrogram", "read_file",
+    "get_spectrogram", "hash_peaks", "plot_peaks", "plot_spectrogram",
+    "generate_waveform", "hash_file", "read_file"
 ]
